+++ liteidex/liteide_fr.ts
@@ -1675,22 +1675,14 @@ Do you want to reload the file from disk?</source>
     </message>
     <message>
         <location filename="src/plugins/golangedit/golangedit.cpp" line="61"/>
+        <location filename="src/plugins/golangedit/golangedit.cpp" line="67"/>
         <source>Find Usages</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="src/plugins/golangedit/golangedit.cpp" line="64"/>
-        <source>Rename Symbol Under Cursor</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="src/plugins/golangedit/golangedit.cpp" line="67"/>
-        <source>Find Usages (Global)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <location filename="src/plugins/golangedit/golangedit.cpp" line="70"/>
-        <source>Rename Symbol Under Cursor (Global)</source>
+        <source>Rename Symbol Under Cursor</source>
         <translation type="unfinished"></translation>
     </message>
 </context>