+++ liteidex/liteide_ja.ts
@@ -1756,23 +1756,15 @@ Answering &quot;Yes&quot; will discard your unsaved changes.</source>
     </message>
     <message>
         <location filename="src/plugins/golangedit/golangedit.cpp" line="61"/>
+        <location filename="src/plugins/golangedit/golangedit.cpp" line="67"/>
         <source>Find Usages</source>
         <translation>使用箇所を検索</translation>
     </message>
     <message>
         <location filename="src/plugins/golangedit/golangedit.cpp" line="64"/>
+        <location filename="src/plugins/golangedit/golangedit.cpp" line="70"/>
         <source>Rename Symbol Under Cursor</source>
         <translation>カーソル下のシンボルをリネーム</translation>
-    </message>
-    <message>
-        <location filename="src/plugins/golangedit/golangedit.cpp" line="67"/>
-        <source>Find Usages (Global)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="src/plugins/golangedit/golangedit.cpp" line="70"/>
-        <source>Rename Symbol Under Cursor (Global)</source>
-        <translation type="unfinished"></translation>
     </message>
     <message>
         <source>Comment/Uncomment Selection</source>