+++ liteidex/src/liteapp/liteapp.cpp
@@ -65,7 +65,7 @@
 #endif
 //lite_memory_check_end
 
-#define LITEIDE_VERSION "X27.2"
+#define LITEIDE_VERSION "X27.2.1"
 
 QString LiteApp::getRootPath()
 {