+++ liteidex/deploy/welcome/zh_CN/changes.md
@@ -11,8 +11,8 @@
 	* fix selection color
 	* fix link color
 * GolangEdit
-	* add find usages (global)
-	* add rename symbol under cursor (global)
+	* add find usages (GOPATH)
+	* add rename symbol under cursor (GOPATH)
 	* fix no jump code tooltip
 	* fix syntax for Non-English symbol name	
 * LiteBuild